requirements.txt
fastapi
uvicorn[standard]
prometheus-client
mcstatus
pyyaml
